% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query.R
\name{build_query}
\alias{build_query}
\title{Build Query}
\usage{
build_query(endpoint, ..., base_url = get_base_url())
}
\arguments{
\item{endpoint}{name of API endpoint. see \code{list_endpoints()}. string}

\item{...}{Arguments for the given endpoint. Use \code{list_endpoint_args(endpoint)} for details. string}

\item{base_url}{the base url of the TOAR REST API. Defaults to the output of \code{get_base_url()}. string}
}
\description{
Constructs a query from the base url, endpoint and relevant arguments
}
