% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint_url.R
\name{get_endpoint_url}
\alias{get_endpoint_url}
\title{Get Endpoint URL}
\usage{
get_endpoint_url(endpoint)
}
\arguments{
\item{endpoint}{name of endpoint. use \code{list_endpoints()} for details. string}
}
\description{
Returns the URL for a named endpoint of the TOAR REST API
}
