% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_endpoints.R
\name{list_controlled_vocabulary}
\alias{list_controlled_vocabulary}
\title{List Controlled Controlled Vocabulary}
\usage{
list_controlled_vocabulary(controlledVocabulary, force = FALSE)
}
\arguments{
\item{force}{force the function to re-scrape the API docs. boolean.}

\item{controlled_vocabulary}{the name of the controlled vocabulary required.
Note these are formatted with begining with each word starting with a capital and
separated with a space. Arguments will be formatted as such using \code{stringr::str_to_title()}.
They will then be formatted for the url via \code{utils:URLencode()}. string.}
}
\description{
Provides the data table associated with the defined controlled Vocabulary.
A description of the controlled vocabulary can be generated with \code{list_endpoint_args("controlled_vocabulary")}
The response will be stored in the internal environment, which can be overridden with \code{force = TRUE}.
}
