% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_endpoints.R
\name{list_endpoint_args}
\alias{list_endpoint_args}
\title{List Endpoint Arguments}
\usage{
list_endpoint_args(endpoint, force = FALSE)
}
\arguments{
\item{endpoint}{name of the API endpoint to return descriptions for. string}

\item{force}{force the function to re-scrape the API docs. boolean}
}
\description{
Returns a tibble containing the given endpoints arguments and their descriptions.
}
\details{
On first run the API docs are scraped and stored in an internal environment.
This can be overridden by setting \code{force = TRUE}.
}
