% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_variables}
\alias{list_variables}
\title{List Variables}
\usage{
list_variables(force = FALSE)
}
\arguments{
\item{force}{force the function to re-scrape the API docs. boolean}
}
\description{
Returns a tibble containing the variables available in the TOAR API
}
