% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_toar_database.R
\name{query_toar_database}
\alias{query_toar_database}
\title{Query TOAR database}
\usage{
query_toar_database(queryUrl)
}
\arguments{
\item{queryUrl}{a query url generated by \code{build_query}. string}
}
\description{
Submits the Query to the TOAR base and handles the response
}
